refactor: Don't recurse when generating a type
This allocates less but most importantly streamlines the logic for the
type module generation.

We now only check once if the type was substituted rather than for each
subslice of the path - it always returned false since it was checking
the *namespace* part of the path but we always specify the full paths in
the type substitution set.